# api/api.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Flio API
host: 'localhost:5000'
basePath: /api/v1
schemes:
  - https
  - http
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /sessions:
    get:
      security:
        - Bearer: []
      tags:
        - Catalogue - Manager
      summary: Lists all sessions created by me
      description: Lists all sessions created by me
      produces:
        - application/json
      responses:
        '200':
          description: Sessions found
          schema:
            type: array
            items:
              $ref: '#/definitions/SessionReply'
        '400':
          description: Invalid query supplied
          schema:
            $ref: '#/definitions/Message'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/Message'
    post:
      security:
        - Bearer: []
      tags:
        - Catalogue - Manager
      summary: Create a session
      description: Create an session
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/SessionBody'
      responses:
        '201':
          description: Feedback created
          schema:
            $ref: '#/definitions/SessionReply'
        '400':
          description: Invalid uuid supplied
          schema:
            $ref: '#/definitions/Message'
        '404':
          description: Session Uuid not found
          schema:
            $ref: '#/definitions/Message'
  '/sessions/{session_uuid}':
    get:
      security:
        - Bearer: []
      tags:
        - Catalogue - Manager
      summary: Get information about the requested session
      description: Get information about the requested session
      parameters:
        - name: session_uuid
          in: path
          required: true
          type: string
          format: uuid
      produces:
        - application/json
      responses:
        '200':
          description: Session found
          schema:
            $ref: '#/definitions/Session'
        '400':
          description: Invalid UUID supplied
          schema:
            $ref: '#/definitions/Message'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/Message'
    put:
      security:
        - Bearer: []
      tags:
        - Catalogue - Manager
      summary: Create a session
      description: Create an session
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: session_uuid
          in: path
          required: true
          type: string
          format: uuid
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/SessionBody'
      responses:
        '204':
          description: Feedback created
          schema:
            $ref: '#/definitions/SessionReply'
        '400':
          description: Invalid uuid supplied
          schema:
            $ref: '#/definitions/Message'
        '404':
          description: Session Uuid not found
          schema:
            $ref: '#/definitions/Message'
    delete:
      security:
        - Bearer: []
      tags:
        - Catalogue - Manager
      summary: Delete a session and all the relative lives / images
      description: A special handling has to be done on the users that buy the lives
      parameters:
        - name: session_uuid
          in: path
          required: true
          type: string
      produces:
        - application/json
      responses:
        '204':
          description: session deleted
          schema:
            $ref: '#/definitions/Message'
        '400':
          description: Invalid UUID supplied
          schema:
            $ref: '#/definitions/Message'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/Message'
...
definitions:
  Message:
    type: string
  LanguageCode:
    type: string
    description: ISO 639-1 language code
    default: en
...
  Category:
    type: object
    required:
      - uuid
      - value
    properties:
      uuid:
        type: string
        format: uuid
        description: UUID of the category. It is the same for all the languages
      value:
        type: string
        example: 'fitness, educational, entertaining, religious'
        description: category in the given language
  Session:
    type: object
    allOf:
      - $ref: '#/definitions/SessionBody'
      - properties:
          uuid:
            type: string
            format: uuid
            description: UUID of the session
          created_at:
            type: string
            format: date-time
            description: >-
              GMT Creation date time of the session as defined by RFC 3339,
              section 5.6
  SessionBody:
    type: object
    required:
      - name
      - category
      - creator_uuid
      - tokens
    properties:
      name:
        type: string
        description: Name of the session
      category:
        $ref: '#/definitions/Category'
      hashtags:
        type: array
        description: Array of hashtags
        items:
          type: string
          example: '#bestof2021'
      description:
        type: string
      language_iso:
        $ref: '#/definitions/LanguageCode'
      creator_uuid:
        type: string
        format: uuid
        description: UUID of the creator (user) of the session
      tokens:
        type: integer
        minimum: 0
        description: Number of tokens needed to attend the session
  SessionReply:
    type: object
    required:
      - session_uuid
    properties:
      session_uuid:
        type: string
        format: uuid
        description: UUID of the session
...